change structure; localStorage-check now in routing.js (routeNerds())
