+++ src/index.html
@@ -63,6 +63,27 @@
         </div> -->
       </section>
 
+      <section data-route="id" id="/src/#stats">
+        <h2> Individual profile </h2>
+        <div id="profile">
+                <article class="student">
+                  <h3 class="projectName"></h3>
+                  <p class="description"></p>
+                  <a class="html_url">
+                    <div class="avatarWrapper">
+                      <img class="avatar" id="avatar">
+                    </div>
+                    <p class="homepage"></p>
+                  </a>
+                  <p class="id"></p>
+                  <p class="node_id"></p>
+                  <p class="name"></p>
+                  <p class="full_name"></p>
+                  <p class="private"></p>
+                </article>
+              </div>
+      </section>
+
     </div>
     <a class="backToTop" href="#top">Back to top</a>
   </main>
